+++ src/components/sign/LoginGroup.tsx
@@ -16,4 +16,22 @@ const LoginGroup = () => {
 
 export default LoginGroup;
 
-const StyledWrapper = styled.div``;
+const StyledWrapper = styled.div`
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+
+  /** 모든 로그인에 공통 적용 */
+
+  .login-img {
+    position: absolute;
+    left: 16px;
+  }
+  .login-text {
+    width: 100%;
+    text-align: center;
+    font-size: 16px;
+    font-weight: bold;
+  }
+`;
