+++ src/pages/App.tsx
@@ -1,23 +1,23 @@
-import React from "react";
-import styled from "styled-components";
-import { BrowserRouter, Route, Routes } from "react-router-dom";
-import Main from "./Main/Main";
-import { useDispatch, useSelector } from "react-redux";
-import { RootState } from "../modules/store";
-import { useEffect } from "react";
-import { loginCheck } from "../modules/store/Login";
-import { useGetUser } from "../hooks/sign/useGetUser";
-import NewSchedule from "./NewSchedule/NewSchedule";
-import MyPage from "./MyPage/MyPage";
-import ScheduleForm from "./ScheduleForm/ScheduleForm";
-import Page404 from "./Page404";
-import ScheduleResult from "./ScheduleResult/ScheduleResult";
-import NavigationBar from "../components/NavigationBar";
+import React from 'react';
+import styled from 'styled-components';
+import { BrowserRouter, Route, Routes } from 'react-router-dom';
+import Main from './Main/Main';
+import { useDispatch, useSelector } from 'react-redux';
+import { RootState } from '../modules/store';
+import { useEffect } from 'react';
+import { loginCheck } from '../modules/store/Login';
+import { useGetUser } from '../hooks/sign/useGetUser';
+import NewSchedule from './NewSchedule/NewSchedule';
+import MyPage from './MyPage/MyPage';
+import ScheduleForm from './ScheduleForm/ScheduleForm';
+import Page404 from './Page404';
+import ScheduleResult from './ScheduleResult/ScheduleResult';
+import NavigationBar from '../components/NavigationBar';
 
 function App() {
   const dispatch = useDispatch();
   const isLogin = useSelector((state: RootState) => state.login.isLogin);
-  const token = localStorage.getItem("access-token");
+  const token = localStorage.getItem('access-token');
   const getUser = useGetUser(token);
 
   useEffect(() => {
@@ -48,8 +48,7 @@ function App() {
 }
 
 const StyledWrapper = styled.div`
-  min-height: 100vh;
-  /* background-color: #d6d6d6; */
+  height: 2000px;
 `;
 
 export default App;
