+++ src/pages/Main/Main.tsx
@@ -1,75 +1,26 @@
-import React from 'react';
+import React, { useEffect } from 'react';
 import styled from 'styled-components';
-import LogoutBtn from '../../components/sign/LogoutBtn';
+import LandingComponent from '../../components/main/LandingComponent';
 
 const Main = () => {
-  const handleSchedule = () => {
-    console.log('일정 생성하기');
+  const handleScroll = () => {
+    console.log('scroll');
   };
+
+  useEffect(() => {
+    window.addEventListener('scroll', handleScroll);
+    return () => {
+      window.removeEventListener('scroll', handleScroll);
+    };
+  }, []);
+
   return (
     <StyledWrapper>
-      {/* <LogoutBtn /> */}
-      <div className="main-wrapper">
-        <div className="main-text">
-          <div>
-            빠르고 간편하게
-            <br /> 지인들과 일정을 잡으세요.
-          </div>
-        </div>
-        <button className="main-btn" onClick={handleSchedule}>
-          일정 생성하기
-        </button>
-      </div>
-      <img src="/img/main.png" className="main-img" alt="main" />
-      <img src="/img/arrow.png" className="main-arrow" alt="arrow" />
+      <LandingComponent />
     </StyledWrapper>
   );
 };
 
 export default Main;
 
-const StyledWrapper = styled.div`
-  height: 100vh;
-  display: flex;
-  justify-content: center;
-  margin-top: 88px;
-  .main-wrapper {
-    z-index: 10;
-    text-align: center;
-    .main-text {
-      font-size: 60px;
-      color: #000070;
-      line-height: 70px;
-      margin-bottom: 50px;
-    }
-    .main-btn {
-      width: 384px;
-      height: 60px;
-      color: white;
-      font-size: 16px;
-      background-color: #000070;
-      border: none;
-      cursor: pointer;
-      border-radius: 8px;
-    }
-  }
-  .main-img {
-    z-index: 1;
-    position: fixed;
-    bottom: -50px;
-    width: 1000px;
-    height: 700px;
-  }
-  .main-arrow {
-    height: 96px;
-    width: 96px;
-    position: absolute;
-    bottom: 10px;
-  }
-  @media screen and (max-width: 1024px) {
-    .main-img {
-      width: 976px;
-      height: 700px;
-    }
-  }
-`;
+const StyledWrapper = styled.div``;
