+++ src/styles/globalStyle.ts
@@ -1,11 +1,8 @@
-import { createGlobalStyle } from "styled-components";
-import { normalize } from "styled-normalize";
+import { createGlobalStyle } from 'styled-components';
+import { normalize } from 'styled-normalize';
 
 const GlobalStyle = createGlobalStyle`
     ${normalize}
-    html,body {
-    overflow: hidden;
-  }
 
   * {
     box-sizing: border-box;
